requirements.txt
aiofiles==23.1.0
celery==5.2.7
fastapi==0.95.0
flower==1.2.0
Jinja2==3.1.2
pytest==7.2.2
redis==4.5.4
requests==2.31.0
uvicorn==0.21.1
httpx==0.23.3
jsonschema==4.17.3
pydantic==1.10.9
python-dotenv==1.0.0
starlette==0.26.1
python-decouple==3.8
loguru==0.7.2
agricore_sp_models>=0.5.85
httpcore==0.16.3
celery[redis]==5.2.7
PyGithub==2.3.0
gurobipy==10.0.0
numpy==1.24.1
wget==3.2
psutil==5.9.8
matplotlib==3.6.3
pandas==1.5.3
dask==2024.5.1
dask[distributed]==2024.5.1
boto3

